+++ lumber-generate.js
@@ -1,6 +1,7 @@
 const P = require('bluebird');
 const program = require('commander');
 const chalk = require('chalk');
+const ora = require('ora');
 const Database = require('./services/database');
 const DatabaseAnalyzer = require('./services/database-analyzer');
 const Dumper = require('./services/dumper');
@@ -8,6 +9,7 @@ const CommandGenerateConfigGetter = require('./services/command-generate-config-
 const logger = require('./services/logger');
 const eventSender = require('./services/event-sender');
 const ProjectCreator = require('./services/project-creator');
+const { terminate } = require('./utils/terminator');
 
 program
   .description('Generate a backend application with an ORM/ODM configured')
@@ -30,31 +32,44 @@ program
 
   const config = await new CommandGenerateConfigGetter(program).perform();
   let schema = {};
+
   if (program.db) {
-    const connection = await new Database().connect(config);
-    schema = await new DatabaseAnalyzer(connection, config, true).perform();
+    const connectionPromise = new Database().connect(config);
+    logger.spinner = ora.promise(connectionPromise, 'Connecting to database');
+    const connection = await connectionPromise;
+
+    const schemaPromise = new DatabaseAnalyzer(connection, config, true).perform();
+    logger.spinner = ora.promise(schemaPromise, 'Analyzing your database');
+    schema = await schemaPromise;
   }
 
-  const { envSecret, authSecret } = await new ProjectCreator(logger)
+  const projectCreationPromise = new ProjectCreator(logger)
     .createProject(config.appName, config);
+  logger.spinner = ora.promise(projectCreationPromise, 'Creating your project on forestadmin');
+  const { envSecret, authSecret } = await projectCreationPromise;
   config.forestEnvSecret = envSecret;
   config.forestAuthSecret = authSecret;
 
+  const spinnerDumper = ora('Creating your project files').start();
+  logger.spinner = spinnerDumper;
   const dumper = await new Dumper(config);
 
   await P.each(Object.keys(schema), async (table) => {
     await dumper.dump(table, schema[table]);
   });
+  spinnerDumper.succeed();
 
   logger.success(`Hooray, ${chalk.green('installation success')}!`);
   await eventSender.notifySuccess();
   process.exit(0);
 })().catch(async (error) => {
-  logger.error(
+  const logs = [
     'Cannot generate your project.',
     'An unexpected error occured. Please create a Github issue with following error:',
-  );
-  logger.log(error);
-  await eventSender.notifyError();
-  process.exit(1);
+    error,
+  ];
+
+  await terminate(1, {
+    logs,
+  });
 });