+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Command generate - Add comments to generated models.
 - Command Generate - Generate the forest collection files.
 - Command Generate - Add tests on model generation (MySQL and PostgreSQL).
+- Command Generate - Add logs about what is happening.
 
 ### Changed
 - Commands - Drop environment variables.